moved common dashboard components to common folder
