+++ app/pages/common/dashboard/components/dashboard-section.njk
