+++ app/pages/common/dashboard/components/dashboard-sectionGroup.njk
