+++ app/pages/supplier/dashboard/subDashboards/template.njk
@@ -1,6 +1,6 @@
 {% extends 'includes/layout.njk' %}
 {% from 'components/back-link/macro.njk' import backLink %}
-{% from '../components/dashboard-sectionGroup.njk' import dashboardSectionGroup %}
+{% from '../../../common/dashboard/components/dashboard-sectionGroup.njk' import dashboardSectionGroup %}
 
 {% block body %}
 <div class="nhsuk-grid-row">
