+++ app/pages/supplier/dashboard/template.njk
@@ -1,6 +1,6 @@
 {% extends 'includes/layout.njk' %}
 {% from 'components/button/macro.njk' import button %}
-{% from './components/dashboard-sectionGroup.njk' import dashboardSectionGroup %}
+{% from '../../common/dashboard/components/dashboard-sectionGroup.njk' import dashboardSectionGroup %}
 {% from '../../../common/components/error-summary/macro.njk' import errorSummary %}
 
 {% block body %}
